split into two libraries: rust-postgres-rest and rust-postgres-rest-actix
