+++ Cargo.toml
@@ -1,48 +1,5 @@
-[package]
-name = "experiment00"
-version = "0.1.0"
-authors = ["Kai Yao <kai.b.yao@gmail.com>"]
-edition = "2018"
-
-[lib]
-name = "experiment00"
-path = "src/lib.rs"
-
-[features]
-default = ["stats_cache"]
-stats_cache = ["clokwerk", "tokio"]
-
-[dependencies]
-actix = "0.8.3"
-actix-web = "1.0.5"
-
-# DO NOT USE BB8. INTEGRATION TESTS HAD DB CONNECTION TIMEOUTS.
-# If you really need a connection pool, look into pgbouncer or L337-postgres.
-# bb8 = "^0.3"
-# bb8-postgres = { version = "^0.3", features = ["with-bit-vec-0_5", "with-chrono-0_4", "with-eui48-0_4", "with-eui48-0_4", "with-serde_json-1", "with-uuid-0_7"] }
-
-chrono = { version = "0.4.7", features = ["serde"] }
-clokwerk = { version = "0.2.1", optional = true }
-eui48 = { version = "0.4.6", features = ["serde"] }
-failure = "0.1.5"
-futures = "0.1.28"
-futures03 = { package = "futures-preview", version = "0.3.0-alpha.17", features = ["compat"] }
-lazy_static = "1.3"
-postgres-protocol = "0.4.1"
-rayon = "1.1.0"
-regex = "1.2.0"
-#rust_decimal = { version = "1.0.1", git = "https://github.com/kaibyao/rust-decimal.git", features = ["postgres"] }
-serde = { version = "1.0.97", features = ["derive"] }
-serde_json = "1.0.40"
-sqlparser = "0.4.0"
-tokio = { version = "0.1.22", optional = true }
-tokio-postgres = { version = "0.4.0-rc.3", features = ["with-chrono-0_4", "with-eui48-0_4", "with-eui48-0_4", "with-serde_json-1", "with-uuid-0_7"] }
-uuid = { version = "0.7.4", features = ["serde"] }
-
-[dependencies.rust_decimal]
-git = "https://github.com/kaibyao/rust-decimal.git"
-features = ["postgres", "serde"]
-
-[dev-dependencies]
-pretty_assertions = "0.6.1"
-reqwest = "0.9.19"
+[workspace]
+members = [
+    "rust-postgres-rest",
+    "rust-postgres-rest-actix",
+]
