+++ rust-postgres-rest-actix/src/endpoints/mod.rs
@@ -5,3 +5,5 @@ mod table;
 pub use table::{
     delete_table, execute_sql, get_all_table_names, get_table, post_table, put_table, reset_caches,
 };
+
+mod query_params_from_request;
