+++ rust-postgres-rest/src/error.rs
@@ -1,14 +1,5 @@
-use actix_web::{http, HttpResponse};
 use failure::Fail;
 use serde::Serialize;
-
-#[derive(Debug, Serialize)]
-/// Describes the type of notification we are sending
-pub enum MessageCategory {
-    Error,
-    /* Info,
-     * Warning, */
-}
 
 #[derive(Debug, Fail, Serialize)]
 #[serde(untagged)]
@@ -20,7 +11,6 @@ pub enum Error {
         code, message, offender, details
     )]
     UserError {
-        category: MessageCategory,
         code: &'static str,
         details: String,
         message: &'static str,
@@ -31,7 +21,6 @@ pub enum Error {
     /// Describes errors that are generated due to system errors.
     #[fail(display = "An internal error has occurred: {}. {}", message, details)]
     InternalError {
-        category: MessageCategory,
         code: &'static str,
         details: String,
         message: &'static str,
@@ -42,7 +31,6 @@ pub enum Error {
 impl From<actix::MailboxError> for Error {
     fn from(err: actix::MailboxError) -> Self {
         Error::InternalError {
-            category: MessageCategory::Error,
             code: "SEND_MESSAGE_ERROR",
             details: format!("{}", err),
             message: "A message failed to send/receive to/from Actix actor.",
@@ -50,32 +38,9 @@ impl From<actix::MailboxError> for Error {
         }
     }
 }
-impl From<actix_web::Error> for Error {
-    fn from(err: actix_web::Error) -> Self {
-        Error::InternalError {
-            category: MessageCategory::Error,
-            code: "ACTIX_ERROR",
-            details: format!("{}", err),
-            message: "Error occurred with Actix.",
-            http_status: 500,
-        }
-    }
-}
-impl From<actix_web::error::PayloadError> for Error {
-    fn from(err: actix_web::error::PayloadError) -> Self {
-        Error::InternalError {
-            category: MessageCategory::Error,
-            code: "PAYLOAD_ERROR",
-            details: format!("{}", err),
-            message: "Could not parse request payload.",
-            http_status: 500,
-        }
-    }
-}
 impl From<chrono::format::ParseError> for Error {
     fn from(err: chrono::format::ParseError) -> Self {
         Error::UserError {
-            category: MessageCategory::Error,
             code: "JSON_ERROR",
             details: format!("{}", err),
             message: "An error occurred when parsing JSON.",
@@ -87,7 +52,6 @@ impl From<chrono::format::ParseError> for Error {
 impl From<eui48::ParseError> for Error {
     fn from(err: eui48::ParseError) -> Self {
         Error::UserError {
-            category: MessageCategory::Error,
             code: "MAC_ADDR_ERROR",
             details: format!("{}", err),
             message: "An error occurred when parsing a mac address.",
@@ -99,7 +63,6 @@ impl From<eui48::ParseError> for Error {
 impl From<rust_decimal::Error> for Error {
     fn from(err: rust_decimal::Error) -> Self {
         Error::UserError {
-            category: MessageCategory::Error,
             code: "DECIMAL_ERROR",
             details: format!("{}", err),
             message: "An error occurred when parsing a decimal string.",
@@ -111,7 +74,6 @@ impl From<rust_decimal::Error> for Error {
 impl From<serde_json::error::Error> for Error {
     fn from(err: serde_json::error::Error) -> Self {
         Error::UserError {
-            category: MessageCategory::Error,
             code: "JSON_ERROR",
             details: format!("{}", err),
             message: "A message occurred when parsing JSON.",
@@ -128,7 +90,6 @@ impl From<sqlparser::parser::ParserError> for Error {
         };
 
         Error::UserError {
-            category: MessageCategory::Error,
             code: "SQL_PARSER_ERROR",
             details,
             message: "A message occurred when parsing SQL.",
@@ -140,7 +101,6 @@ impl From<sqlparser::parser::ParserError> for Error {
 impl<T> From<std::sync::PoisonError<T>> for Error {
     fn from(err: std::sync::PoisonError<T>) -> Self {
         Error::InternalError {
-            category: MessageCategory::Error,
             code: "MEM_LOCK_ERROR",
             details: format!("{}", err),
             message: "A memory-locked process has failed.",
@@ -151,7 +111,6 @@ impl<T> From<std::sync::PoisonError<T>> for Error {
 impl From<tokio_postgres::Error> for Error {
     fn from(err: tokio_postgres::Error) -> Self {
         Error::InternalError {
-            category: MessageCategory::Error,
             code: "DATABASE_ERROR",
             details: format!("{}", err),
             message: "A database error occurred (postgres).",
@@ -162,7 +121,6 @@ impl From<tokio_postgres::Error> for Error {
 impl From<uuid::parser::ParseError> for Error {
     fn from(err: uuid::parser::ParseError) -> Self {
         Error::UserError {
-            category: MessageCategory::Error,
             code: "UUID_ERROR",
             details: format!("{}", err),
             message: "An error occurred when parsing a UUID string.",
@@ -186,7 +144,6 @@ impl Error {
     pub fn generate_error(err_id: &'static str, offender: String) -> Self {
         match err_id {
             "INCORRECT_REQUEST_BODY" => Error::UserError {
-                category: MessageCategory::Error,
                 code: err_id,
                 details: "".to_string(),
                 http_status: 400,
@@ -195,7 +152,6 @@ impl Error {
             },
 
             "INVALID_CONTENT_TYPE" => Error::UserError {
-                category: MessageCategory::Error,
                 code: err_id,
                 details: "The `Content-Type` header value is not valid for this request".to_string(),
                 http_status: 400,
@@ -204,7 +160,6 @@ impl Error {
             },
 
             "INVALID_DB_CONFIG" => Error::UserError {
-                category: MessageCategory::Error,
                 code: err_id,
                 details: "Either `db_url` or `db_pool` config property needs to be set.".to_string(),
                 http_status: 400,
@@ -213,7 +168,6 @@ impl Error {
             },
 
             "INVALID_JSON_TYPE_CONVERSION" => Error::UserError {
-                category: MessageCategory::Error,
                 code: err_id,
                 details: "The type of the JSON data does not match the type of the database column.".to_string(),
                 http_status: 400,
@@ -221,17 +175,7 @@ impl Error {
                 offender
             },
 
-            "INVALID_PREPARED_VALUE_TYPE_CONVERSION" => Error::UserError {
-                category: MessageCategory::Error,
-                code: err_id,
-                details: ".".to_string(),
-                http_status: 400,
-                message: "Could not convert ParsedSQLValue type to TypedColumnValue.",
-                offender
-            },
-
             "INVALID_SQL_IDENTIFIER" => Error::UserError {
-                category: MessageCategory::Error,
                 code: err_id,
                 details: "Valid identifiers must only contain alphanumeric and underscore (_) characters. The first character must also be a letter or underscore. Wildcards (*) are not allowed.".to_string(),
                 http_status: 400,
@@ -240,7 +184,6 @@ impl Error {
             },
 
             "INVALID_SQL_SYNTAX" => Error::UserError {
-                category: MessageCategory::Error,
                 code: err_id,
                 details: "The SQL expression could not be parsed by PostgreSQL.".to_string(),
                 http_status: 400,
@@ -248,26 +191,7 @@ impl Error {
                 offender,
             },
 
-            "NO_DATABASE_CONNECTION" => Error::UserError {
-                category: MessageCategory::Error,
-                code: err_id,
-                details: "A database client does not exist.".to_string(),
-                http_status: 500,
-                message: "Something went wrong during server startup. Message the admin.",
-                offender,
-            },
-
-            "REQUIRED_PARAMETER_MISSING" => Error::UserError {
-                category: MessageCategory::Error,
-                code: err_id,
-                details: "".to_string(),
-                http_status: 400,
-                message: "There was a parameter required by this action, but it was not found.",
-                offender,
-            },
-
             "SQL_IDENTIFIER_KEYWORD" => Error::UserError {
-                category: MessageCategory::Error,
                 code: err_id,
                 details: "`table` is a reserved keyword and cannot be used to name SQL identifiers".to_string(),
                 http_status: 400,
@@ -276,15 +200,21 @@ impl Error {
             },
 
             "TABLE_COLUMN_TYPE_NOT_FOUND" => Error::InternalError {
-                category: MessageCategory::Error,
                 code: err_id,
                 details: format!("The column type for column `{}` could not be generated from the Table Stats query. Please submit a bug report, as this really shouldn’t be happening.", offender),
                 http_status: 500,
                 message: "The column type for a queried table column could not be determined.",
             },
 
+            "TABLE_STATS_CACHE_NOT_ENABLED" => Error::UserError {
+                code: err_id,
+                details: "The Table Stats Cache has not been enabled, please turn it on by calling the config’s `cache_table_stats()`.".to_string(),
+                http_status: 400,
+                message: "",
+                offender,
+            },
+
             "TABLE_STATS_CACHE_NOT_INITIALIZED" => Error::UserError {
-                category: MessageCategory::Error,
                 code: err_id,
                 details: "The Table Stats Cache has not yet started/finished fetching table stats.".to_string(),
                 http_status: 503,
@@ -293,7 +223,6 @@ impl Error {
             },
 
             "UNSUPPORTED_DATA_TYPE" => Error::UserError {
-                category: MessageCategory::Error,
                 code: err_id,
                 details: "".to_string(),
                 http_status: 400,
@@ -303,60 +232,12 @@ impl Error {
 
             // If this happens, that means we forgot to implement an error handler
             _ => Error::UserError {
-                category: MessageCategory::Error,
                 code: err_id,
                 details: "Generic error.".to_string(),
                 http_status: 418,
                 message: "An error occurred that we did not anticipate. Please let admins know.",
                 offender,
             }
-        }
-    }
-}
-
-/// Used for formatting the Errors that occur to display in an http response.
-#[derive(Debug, Serialize)]
-struct DisplayUserError<'a> {
-    code: &'static str,
-    details: String,
-    message: &'static str,
-    offender: Option<&'a str>,
-}
-
-// How Errors are formatted for an http response
-impl actix_web::ResponseError for Error {
-    fn render_response(&self) -> HttpResponse {
-        match self {
-            Error::UserError {
-                code,
-                details,
-                http_status,
-                message,
-                offender,
-                ..
-            } => HttpResponse::build(http::StatusCode::from_u16(*http_status).unwrap()).json(
-                DisplayUserError {
-                    code,
-                    details: details.to_string(),
-                    message,
-                    offender: Some(offender),
-                },
-            ),
-
-            Error::InternalError {
-                code,
-                details,
-                http_status,
-                message,
-                ..
-            } => HttpResponse::build(http::StatusCode::from_u16(*http_status).unwrap()).json(
-                DisplayUserError {
-                    code,
-                    details: details.to_string(),
-                    message,
-                    offender: None,
-                },
-            ),
         }
     }
 }
