+++ rust-postgres-rest/src/stats_cache.rs
@@ -1,7 +1,7 @@
 use crate::{
     db::connect,
     queries::{select_all_table_stats, select_all_tables, TableStats},
-    AppConfig, AppState, Error,
+    Config, Error,
 };
 use actix::{spawn, Actor, Addr, Context, Handler, Message, ResponseFuture};
 use clokwerk::{Scheduler, TimeUnits};
@@ -92,7 +92,7 @@ impl Handler<StatsCacheMessage> for StatsCache {
 
 impl StatsCache {
     /// Creates a new instance of `StatsCache`.
-    pub fn new(config: AppConfig) -> Self {
+    pub fn new(config: Config) -> Self {
         StatsCache {
             db_url: config.db_url,
             cache: Arc::new(RwLock::new(None)),
@@ -123,8 +123,7 @@ impl StatsCache {
             let cache_clone = self.cache.clone();
 
             let f = connect(self.db_url)
-                .and_then(select_all_tables)
-                .map_err(Error::from)
+                .and_then(|client| select_all_tables(client).map_err(Error::from))
                 .and_then(|(tables, client)| {
                     select_all_table_stats(client, tables).boxed().compat()
                 })
@@ -152,28 +151,29 @@ lazy_static! {
     static ref STATS_CACHE_MUTEX: Mutex<Option<Addr<StatsCache>>> = Mutex::new(None);
 }
 
-pub fn initialize_stats_cache(state: &mut AppState) {
+pub fn initialize_stats_cache(config: &mut Config) {
     // initialize table stats cache actor once, then clone its address for use across threads
     if IS_STATS_CACHE_INIT.load(Ordering::SeqCst) {
         let static_addr_opt = STATS_CACHE_MUTEX.lock().unwrap();
-        state.stats_cache_addr = Some(static_addr_opt.as_ref().unwrap().clone());
+        config.stats_cache_addr = Some(static_addr_opt.as_ref().unwrap().clone());
     } else {
-        let config = state.config.clone();
+        let config_clone = config.clone();
+        let cache_reset_interval_seconds = config.cache_reset_interval_seconds;
         INIT_STATS_CACHE.call_once(|| {
             let mut static_addr_opt = STATS_CACHE_MUTEX.lock().unwrap();
-            let addr = StatsCache::new(config).start();
+            let addr = StatsCache::new(config_clone).start();
             let addr_clone = addr.clone();
             let addr_clone_2 = addr.clone();
             static_addr_opt.get_or_insert(addr);
 
-            state.stats_cache_addr = Some(addr_clone);
+            config.stats_cache_addr = Some(addr_clone);
             IS_STATS_CACHE_INIT.store(true, Ordering::SeqCst);
 
             // set cache reset interval
-            if state.config.cache_reset_interval_seconds > 0 {
+            if cache_reset_interval_seconds > 0 {
                 let mut scheduler = Scheduler::new();
                 scheduler
-                    .every(state.config.cache_reset_interval_seconds.seconds())
+                    .every(cache_reset_interval_seconds.seconds())
                     .run(move || {
                         let reset_future =
                             addr_clone_2
